requirements.txt
python-telegram-bot==20.0
nest-asyncio==1.5.6  # Required for async handling

